+++ src/App.js
@@ -15,16 +15,16 @@ const store = createStore(rootReducer);
 function App() {
   // TODO: Link the query params to the correct IV level
   // TODO: If the query params do not belong to any condition number or the trial task is not a value from 1-3 inclusive, do not allow the user to proceed (show "Invalid Link")
-  const [searchParams] = useSearchParams();
-  const a = searchParams.get("a");
-  const b = searchParams.get("b");
-  const isInit = loggingjs.isInitialized();
-  const path = useLocation().pathname;
-  const isRoot = path === "/";
-  const isValidLink = (a && b && isRoot) || isInit;
-  if (!isValidLink) {
-    return <div>Invalid Link</div>;
-  }
+  // const [searchParams] = useSearchParams();
+  // const a = searchParams.get("a");
+  // const b = searchParams.get("b");
+  // const isInit = loggingjs.isInitialized();
+  // const path = useLocation().pathname;
+  // const isRoot = path === "/";
+  // const isValidLink = (a && b && isRoot) || isInit;
+  // if (!isValidLink) {
+  //   return <div>Invalid Link</div>;
+  // }
 
   return (
     <Provider store={store}>
