+++ src/components/ModuleCart.module.css
@@ -1,0 +1,44 @@
+/* .moduleCart {
+  position: sticky;
+  right: 0;
+  width: 200px;
+  height: 100px;
+  box-sizing: border-box;
+  background: #f9f9f9;
+  border: 2px solid #cdd1d7;
+  border-radius: 10px 10px 0px 0px;
+  display: flex;
+  align-items: center;
+  justify-content: flex-start;
+  padding: 0 30px;
+  overflow-y: auto;
+  word-wrap: normal;
+}
+*/
+
+.moduleCart {
+  position: sticky;
+  width: 250px;
+  height: 500px;
+  box-sizing: border-box;
+  background: #f9f9f9;
+  border: 2px solid #cdd1d7;
+  border-radius: 10px;
+  display: flex;
+  flex-direction: column;
+  align-items: flex-start;
+  justify-content: flex-start;
+  padding: 25px;
+  overflow-y: auto;
+  margin-left: 180px;
+  margin-top: 50px;
+}
+
+.title {
+  font-weight: bold;
+}
+
+.item {
+  color: #7d9ab4;
+  margin: 0;
+}
