+++ src/components/Page.module.css
@@ -32,6 +32,15 @@
   flex: 1;
   overflow-y: auto !important;
   position: relative;
+  text-align: left;
+}
+
+.left {
+  display: inline-block;
+}
+
+.right {
+  display: inline-block;
 }
 
 .content {
@@ -54,10 +63,10 @@
   font-size: 15px;
   vertical-align: middle;
   box-shadow: inset 0 1px 0 0 #74a7cb;
-  padding: 0.3em .7em;
+  padding: 0.3em 0.7em;
   border: 1px solid #1c3f5d;
   border-bottom-color: #0d2940;
-  background-image: linear-gradient(rgb(054,102,139),rgb(027,060,088));
+  background-image: linear-gradient(rgb(054, 102, 139), rgb(027, 060, 088));
   text-align: left;
   display: inline-block;
   background-color: transparent;
@@ -79,7 +88,7 @@
 }
 
 .whitebutton {
-  padding: 0.3em .7em;
+  padding: 0.3em 0.7em;
   height: 35px;
   width: 35px;
   border: none;
